Move mutation over to assignment
