+++ data/schema.graphql
@@ -254,7 +254,7 @@ type RootMutationType {
   createReview(courseId: ID!, grade: Int, revisionRequested: Boolean!, submissionId: ID!): SubmissionType!
 
   """Creates a new submission for the viewer"""
-  createSubmission(assignmentId: ID!, courseId: ID!, pullRequestUrl: String!): SubmissionType!
+  createSubmission(assignmentId: ID!, courseId: ID!, pullRequestUrl: String!): AssignmentType
 
   """Generates an invitation code"""
   generateInviteCode(courseId: ID!, expirationMinutes: Int, roleId: ID!): String!
